+++ sdk/compute/Azure.ResourceManager.Compute/src/Customize/CloudService/Models/CloudServiceVaultAndSecretReference.cs
@@ -25,6 +25,7 @@ namespace Azure.ResourceManager.Compute.Models
         /// <summary> The secret URI. </summary>
         public Uri SecretUri { get; set; }
 
+        /// <summary> Writes the JSON representation of the model to the provided writer. </summary>
         protected virtual void JsonModelWriteCore(Utf8JsonWriter writer, ModelReaderWriterOptions options)
             => throw new NotSupportedException("CloudService operations are no longer supported.");